Add an x86 JIT for v2.
The stack now grows upward, so old tricks for variadic arguments get
more complicated. Instead, we now box variadic arguments into a local
array, and pass the array to target functions. This is entirely handled
in the lowerer.

We have also introduced a new error reporting model for fast calls to
C++.  Error reports can now be deferred so exit frames aren't needed.

+++ vm/x64/assembler-x64.h
@@ -427,7 +427,7 @@ class Assembler : public AssemblerBase
 
             uint32_t data_index;
             if (src_pc >= 0) {
-                assert(src_pc <= code_size());
+                assert(static_cast<size_t>(src_pc) <= code_size());
 
                 uint8_t* base = buffer() + src_pc - 6;
                 assert(*base == 0xff);
@@ -590,7 +590,7 @@ class Assembler : public AssemblerBase
 
         intptr_t rip = pc();
         if (src->bound()) {
-            assert(src->offset() < pc());
+            assert(static_cast<uint32_t>(src->offset()) < pc());
 
             intptr_t delta = rip - intptr_t(src->offset());
             assert(delta >= INT_MIN && delta <= INT_MAX);
